server/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
blinker==1.9.0
click==8.2.1
fastapi==0.116.1
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
h11==0.16.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
psycopg2==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.47.3
typing-inspection==0.4.1
typing_extensions==4.15.0
uvicorn==0.35.0
Werkzeug==3.1.3
